Refined docker image build process
Changed how the primary docker image is built by using a secondary docker
image. This has been done primarily to speed up the build process under
GitHub Actions as we build for `linux/amd64`, `linux/arm64`, and
`linux/arm/v7`, and the construction of the pgTAP scripts did not need to be
repeated three times (there is no difference based on the architecture, just
on the PostgreSQL version).

This change means that we can support more versions of PostgreSQL without
caring about released base image as there are no compiled files. As such, we
have re-added support for PostgreSQL 9.6 on a best-effort basis.

- Added a Maintenance document.

- Docker build arguments are no longer defaulted in the main `Dockerfile`,
  but are stored in `package-versions.json`. The `Justfile` reads these for
  use with the `build` target (`just build`) and the GitHub Actions workflow
  via `fromJson`.